feat: stimulus node support in builder, unified canvas primitives, and UX fixes
- Add stimulus node to ensemble builder UI with config panel, drag-to-target
  wiring, and "Add Stimulus" toolbar button
- Unify canvas components: extract shared node types (AgentConfigNode,
  StimulusNode, ProviderNode, ModelNode), edge styles, layout helpers, and
  shell components into canvas-primitives.tsx — single source of truth for
  both builder and read-only canvases
- Add default stimulus configs to 5 default ensembles (developer-team,
  research-team, platform-team, devops-essentials, observability-mcp-team)
- Include stimulus field in API create payload
- Fix otel collector DNS: use FQDN (.svc.cluster.local) and make URL
  configurable via SYMPOZIUM_COLLECTOR_METRICS_URL env var in Helm chart
- Add ensemble delete button with confirmation dialog to ensembles list page
- Improve collector status display: show "Not reachable" with helpful hint
  instead of raw DNS error
- Fix pre-existing bug: edge type filter used e.data.type instead of
  e.data.relType in builder settings panel
- Fix integration test: use -lt instead of -ne for default ensemble count
