+++ charts/sympozium/files/agent-configs/developer-team.yaml
@@ -13,6 +13,9 @@ spec:
   version: "1.0.0"
   workflowType: delegation
   relationships:
+    - source: kickoff
+      target: tech-lead
+      type: stimulus
     - source: tech-lead
       target: backend-dev
       type: delegation
@@ -37,6 +40,9 @@ spec:
     - source: code-reviewer
       target: frontend-dev
       type: supervision
+  stimulus:
+    name: kickoff
+    prompt: "Review the repository for open issues and pull requests. Triage unassigned issues, review open PRs, run tests, and submit worthwhile pull requests to improve the project. Focus on code quality, test coverage, and documentation gaps."
   sharedMemory:
     enabled: true
     storageSize: "1Gi"
@@ -61,6 +67,13 @@ spec:
         delegate to the appropriate team member by creating well-scoped issues.
         When you review code, you focus on correctness, architecture, and
         maintainability over style.
+
+        When a task requires implementation, use the delegate_to_persona tool
+        to delegate to the right team member:
+        - "backend-dev" for server-side features, APIs, and business logic
+        - "frontend-dev" for UI components, styling, and client-side logic
+        - "devops-engineer" for CI/CD, infrastructure, and deployment config
+        Include clear context and acceptance criteria in every delegation.
       skills:
         - github-gitops
         - code-review
