+++ charts/sympozium/files/agent-configs/observability-mcp-team.yaml
@@ -12,7 +12,13 @@ spec:
   category: observability
   version: "1.0.0"
   workflowType: delegation
+  stimulus:
+    name: observe
+    prompt: "Investigate current observability data for anomalies, latency spikes, and error rate changes. Query Dynatrace for new problems, correlate with Kubernetes events and pod status, and escalate any confirmed incidents."
   relationships:
+    - source: observe
+      target: sre-investigator
+      type: stimulus
     - source: sre-investigator
       target: incident-responder
       type: delegation
@@ -34,6 +40,11 @@ spec:
         2. Use dynatrace_get_metrics to check resource utilisation trends
         3. Correlate findings with k8s pod status and events
         4. Produce a structured investigation report with timeline and RCA
+
+        When your investigation confirms an active incident requiring
+        network diagnostics or remediation, use the delegate_to_persona
+        tool to delegate to "incident-responder" with the incident details
+        and your findings so far.
       skills:
         - mcp-bridge
         - k8s-ops
