+++ charts/sympozium/files/agent-configs/research-team.yaml
@@ -14,7 +14,14 @@ spec:
   version: "1.0.0"
   workflowType: delegation
 
+  stimulus:
+    name: research-brief
+    prompt: "Begin a new research cycle. Identify trending topics or open research questions relevant to the project, gather sources, and produce a comprehensive report with citations. Submit findings for review."
+
   relationships:
+    - source: research-brief
+      target: lead
+      type: stimulus
     - source: researcher
       target: writer
       type: delegation
@@ -38,6 +45,19 @@ spec:
       target: reviewer
       type: supervision
 
+  sharedMemory:
+    enabled: true
+    storageSize: "1Gi"
+    accessRules:
+      - agentConfig: lead
+        access: read-write
+      - agentConfig: researcher
+        access: read-write
+      - agentConfig: writer
+        access: read-write
+      - agentConfig: reviewer
+        access: read-only
+
   agentConfigs:
     - name: lead
       displayName: "Research Lead"
@@ -47,7 +67,8 @@ spec:
 
         Your responsibilities:
         - Break down complex research questions into actionable tasks
-        - Delegate research tasks to the Researcher
+        - Delegate research tasks to the Researcher using the
+          delegate_to_persona tool with target "researcher"
         - Monitor the Writer and Reviewer for progress
         - Synthesise final outputs and deliver to stakeholders
         - Maintain a research backlog of pending topics
@@ -79,7 +100,8 @@ spec:
         Your responsibilities:
         - Investigate assigned research topics using web search and file analysis
         - Produce structured research notes with key findings, sources, and data
-        - Delegate to the Writer when your research notes are ready
+        - When your research notes are ready, use the delegate_to_persona
+          tool to delegate to "writer" with the complete research notes
         - Flag any blockers or questions back to the Lead
       skills:
         - memory
