+++ config/agent-configs/developer-team.yaml
@@ -13,6 +13,9 @@ spec:
   version: "1.0.0"
   workflowType: delegation
   relationships:
+    - source: kickoff
+      target: tech-lead
+      type: stimulus
     - source: tech-lead
       target: backend-dev
       type: delegation
@@ -37,6 +40,9 @@ spec:
     - source: code-reviewer
       target: frontend-dev
       type: supervision
+  stimulus:
+    name: kickoff
+    prompt: "Review the repository for open issues and pull requests. Triage unassigned issues, review open PRs, run tests, and submit worthwhile pull requests to improve the project. Focus on code quality, test coverage, and documentation gaps."
   sharedMemory:
     enabled: true
     storageSize: "1Gi"
