+++ config/agent-configs/platform-team.yaml
@@ -11,7 +11,13 @@ spec:
   category: platform
   version: "1.0.0"
   workflowType: delegation
+  stimulus:
+    name: platform-audit
+    prompt: "Audit the cluster for security misconfigurations, resource pressure, and failing workloads. Review namespace configurations against golden path standards and flag any workloads running with excessive privileges or missing resource limits."
   relationships:
+    - source: platform-audit
+      target: platform-engineer
+      type: stimulus
     - source: platform-engineer
       target: security-guardian
       type: delegation
